Change encoding of terminal fonts to just be colors.
The terminal colors now match those indicated by Nethack, with and
additional sign indicating the 'reverse feature'. The fonts can now be
represented with a int8 instead of a short. These should make learning
easier, as users dont have to unpack the bitset.

+++ nle/nethack/nethack.py
@@ -42,7 +42,7 @@ OBSERVATION_DESC = {
     "inv_strs": dict(shape=INV_STRS_SHAPE, dtype=np.uint8),
     "screen_descriptions": dict(shape=SCREEN_DESCRIPTIONS_SHAPE, dtype=np.uint8),
     "tty_chars": dict(shape=TERMINAL_SHAPE, dtype=np.uint8),
-    "tty_colors": dict(shape=TERMINAL_SHAPE, dtype=np.int16),
+    "tty_colors": dict(shape=TERMINAL_SHAPE, dtype=np.int8),
 }
